# Conflicts:
#	web/src/Web.App/Constants/TrackedLinks.cs

+++ platform/src/apis/Platform.Api.Benchmark/CustomData/CustomDataRequest.cs
@@ -42,7 +42,7 @@ public record CustomDataRequest
     public decimal? TotalPupils { get; set; }
     public decimal? PercentFreeSchoolMeals { get; set; }
     public decimal? PercentSpecialEducationNeeds { get; set; }
-    public int? TotalInternalFloorArea { get; set; }
+    public decimal? TotalInternalFloorArea { get; set; }
     public decimal? WorkforceFTE { get; set; }
     public decimal? TeachersFTE { get; set; }
     public decimal? PercentTeacherWithQualifiedStatus { get; set; }
